requirements.txt
python-telegram-bot==21.4
deep-translator==1.11.4
python-dotenv==1.0.1